@@ -1,28 +1,26 @@
-module github.com/jptosso/coraza-benchmarks
+module github.com/corazawaf/coraza-benchmark
 
-go 1.17
+go 1.22
 
 require (
-	github.com/jptosso/coraza-libinjection v0.0.0-20220109185528-78d7edbd2d9f
-	github.com/jptosso/coraza-pcre v0.0.0-20211201204829-1bea0f044896
-	github.com/jptosso/coraza-waf/v2 v2.0.0-rc.1
-	github.com/shirou/gopsutil/v3 v3.21.12
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/corazawaf/coraza/v2 v2.0.1
+	github.com/corazawaf/libinjection-go v0.2.1
+	github.com/gijsbers/go-pcre v0.0.0-20161214203829-a84f3096ab3c
+	github.com/shirou/gopsutil/v3 v3.24.5
+	go.uber.org/zap v1.27.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
-require github.com/gijsbers/go-pcre v0.0.0-20161214203829-a84f3096ab3c // indirect
-
 require (
-	github.com/cloudflare/ahocorasick v0.0.0-20210425175752-730270c3e184 // indirect
-	github.com/go-ole/go-ole v1.2.6 // indirect
-	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
-	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
-	github.com/tklauser/go-sysconf v0.3.9 // indirect
-	github.com/tklauser/numcpus v0.3.0 // indirect
-	github.com/yusufpapurcu/wmi v1.2.2 // indirect
-	go.uber.org/atomic v1.9.0 // indirect
-	go.uber.org/multierr v1.7.0 // indirect
-	go.uber.org/zap v1.19.1 // indirect
-	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4 // indirect
-	golang.org/x/sys v0.0.0-20211013075003-97ac67df715c // indirect
+	github.com/go-ole/go-ole v1.3.0 // indirect
+	github.com/lufia/plan9stats v0.0.0-20240513124658-fba389f38bae // indirect
+	github.com/petar-dambovaliev/aho-corasick v0.0.0-20240411101913-e07a1f0e8eb4 // indirect
+	github.com/power-devops/perfstat v0.0.0-20240221224432-82ca36839d55 // indirect
+	github.com/shoenig/go-m1cpu v0.1.6 // indirect
+	github.com/tklauser/go-sysconf v0.3.14 // indirect
+	github.com/tklauser/numcpus v0.8.0 // indirect
+	github.com/yusufpapurcu/wmi v1.2.4 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
+	golang.org/x/net v0.26.0 // indirect
+	golang.org/x/sys v0.21.0 // indirect
 )
